14 спринт 1  итерация

+++ controllers/cards.js
@@ -1,5 +1,5 @@
 const Card = require('../models/card');
-const { BAD_REQUEST, NOT_FOUND, INTERNAL_SERVER_ERROR } = require('../utils/constants');
+const { BAD_REQUEST, NOT_FOUND, FORBIDDEN, INTERNAL_SERVER_ERROR } = require('../utils/constants');
 
 const getCards = (req, res) => {
   Card.find({})
@@ -23,13 +23,25 @@ const postCards = (req, res) => {
 
 const deleteCards = (req, res) => {
   const { cardId } = req.params;
+  console.log(req.params);
+  const userId = req.user._id;
 
-  Card.findByIdAndRemove(cardId)
+
+  Card.findById(cardId)
     .then((card) => {
       if (!card) {
         return res.status(NOT_FOUND).send({ message: 'Такой карточки не существует' });
       }
-      return res.send(card);
+
+      // Проверка, является ли текущий пользователь владельцем карточки
+      if (card.owner.toString() !== userId) {
+        return res.status(FORBIDDEN).send({ message: 'Доступ запрещен' });
+      }
+
+      // Удаление карточки
+      Card.findByIdAndRemove(cardId)
+        .then((deletedCard) => res.send(deletedCard))
+        .catch(() => res.status(INTERNAL_SERVER_ERROR).send({ message: 'Произошла ошибка' }));
     })
     .catch((err) => {
       if (err.name === 'CastError') {
